+++ packages/backend-openapi-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/backend-openapi-utils
 
+## 0.1.3-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.10-next.2
+  - @backstage/errors@1.2.3
+
 ## 0.1.3-next.1
 
 ### Patch Changes