+++ packages/e2e-test/CHANGELOG.md
@@ -1,5 +1,14 @@
 # e2e-test
 
+## 0.2.12-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/create-app@0.5.11-next.2
+  - @backstage/cli-common@0.1.13
+  - @backstage/errors@1.2.3
+
 ## 0.2.12-next.1
 
 ### Patch Changes