+++ plugins/app-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-app-node
 
+## 0.1.10-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.10-next.2
+
 ## 0.1.10-next.1
 
 ### Patch Changes