+++ plugins/home-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-home-react
 
+## 0.1.8-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.14.0-next.1
+  - @backstage/core-plugin-api@1.9.0-next.1
+
 ## 0.1.8-next.1
 
 ### Patch Changes