+++ plugins/org-react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-org-react
 
+## 0.1.19-next.2
+
+### Patch Changes
+
+- 8fe56a8: Widen `@types/react` dependency range to include version 18.
+- Updated dependencies
+  - @backstage/core-components@0.14.0-next.1
+  - @backstage/core-plugin-api@1.9.0-next.1
+  - @backstage/plugin-catalog-react@1.10.0-next.2
+  - @backstage/catalog-client@1.6.0-next.1
+  - @backstage/catalog-model@1.4.4-next.0
+
 ## 0.1.19-next.1
 
 ### Patch Changes