+++ plugins/scaffolder-backend-module-gerrit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-scaffolder-backend-module-gerrit
 
+## 0.1.2-next.2
+
+### Patch Changes
+
+- 8b269e1: Provide default branch when creating repositories.
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.10-next.2
+  - @backstage/plugin-scaffolder-node@0.3.0-next.2
+  - @backstage/config@1.1.1
+  - @backstage/errors@1.2.3
+  - @backstage/integration@1.9.0-next.0
+
 ## 0.1.2-next.1
 
 ### Patch Changes